iconography page style tweaks
