+++ src/components/markdown/markdown.scss
@@ -97,7 +97,7 @@ h2, h3, h4, h5 {
   &:before {
     display: block; 
     content: " "; 
-    margin-top: -rem(56px); 
+    margin-top: rem(-56px); 
     height: rem(56px); 
     visibility: hidden; 
     pointer-events: none;
