--- src/content/guidelines/iconography-ARCHIVED/library.md
@@ -1,7 +1,0 @@
----
-label: Guidelines
-title: Iconography
-tabs: ['Library', 'Usage', 'Contribution']
----
-
-<icon-library />
