+++ src/content/guidelines/iconography/library.md
@@ -4,7 +4,4 @@ title: Iconography
 tabs: ['Library', 'Usage', 'Design and Contribution']
 ---
 
-<p>The source file for the icon library is the <a href="https://github.com/ibm/carbon-design-kit" target=blank>Carbon Design Kit.</a></p>
-<p> <a href="https://github.com/ibm/carbon-icons" target=blank>Carbon icons GitHub repo</a></p>
-
-<icon-library-experimental />
+<icon-library />
