+++ src/styles/experimental.scss
@@ -19,19 +19,3 @@ $feature-flags: (
   );
   @import 'carbon-components/src/globals/scss/styles';
 }
-
-.iconography--experimental .icon-container {
-  width: 100%;
-}
-
-.iconography--experimental .icon-size {
-  margin-top: 4rem;
-}
-
-.iconography--experimental .icon__card-title,
-.iconography--experimental .icon__card-details {
-  display: block;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
